fix(client): Restructuration dossier styles/pages + fix couleur en thème dark
