+++ .gitignore
@@ -49,4 +49,5 @@ thumb
 sketch
 
 package-lock.json
+yarn.lock
