+++ packages/client/src/pages/Homepage.css
@@ -15,6 +15,10 @@
   margin-bottom: 10px;
 }
 
+.dark-theme .hero-section h1 {
+	color: #f0f0f0;
+}
+
 .tagline {
   font-size: 1.2rem;
   color: #666;
