+++ packages/client/src/pages/Login.jsx
@@ -1,6 +1,6 @@
 import {useState} from 'react';
 import {Link, useNavigate} from 'react-router-dom';
-import '../styles/Login.css';
+import './Login.css';
 
 function Login() {
 	const [formData, setFormData] = useState({
