+++ packages/client/src/pages/Register.jsx
@@ -1,6 +1,6 @@
 import {useState} from 'react';
 import {Link, useNavigate} from 'react-router-dom';
-import '../styles/Register.css';
+import './Register.css';
 
 function Register() {
 	const [formData, setFormData] = useState({
